requirements.txt
Flask==2.3.2
click==8.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
Werkzeug==2.3.4
mysql-connector-python==8.0.33
